requirements.txt
django~=1.11                   # Django 1.11 is designated as a long-term support release
djangorestframework
djangorestframework-jwt
pymysql
markdown
httpie

